requirements.txt
google-genai>=1.0.0
faster-whisper>=1.0.0
pyannote.audio>=3.1.0
speechbrain>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0
edge-tts>=6.1.0
numpy>=1.24.0
scipy>=1.10.0
sounddevice>=0.4.6
faiss-cpu>=1.7.4
aiosqlite>=0.19.0
fastapi>=0.110.0
uvicorn>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
jinja2>=3.1.0
